% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSRF_Version_1_0.R
\name{Complete_DS}
\alias{Complete_DS}
\title{Making Final Data set with non-functional covariate, random split functional covariates and response values.}
\usage{
Complete_DS(
  ResponseVar,
  Covariates,
  RawData,
  Stat,
  min,
  max,
  Params = c(0, 0, 0),
  Distribution,
  Params2 = c(0, 0, 0),
  Distribution2,
  type
)
}
\arguments{
\item{ResponseVar}{The response variable (Continuous, Categorical or Survival )}

\item{Covariates}{The non-functional Covariates.}

\item{RawData}{The functional data that defines in the domain.}

\item{Stat}{The summary statistics for each split including "mean","median","Q25","Q75","sd","max","min","range".}

\item{min}{The minimum value of the domain.}

\item{max}{The maximum value of the domain.}

\item{Params}{The parameters of the distribution. 
In the "Exponential" is (rate, 0,0)
In the "Uniform" is  (min,max,0)
In the "Normal"  is  (Mu, Sigma,0)}

\item{Distribution}{The Distribution name. ("Exponential", "Uniform" and "Normal").}

\item{Params2}{The parameters of the second distribution, if type is "Overlap".
In the "Exponential" is (rate, Constant,0), Constant default is 5. Final rate is rate*Constant.
In the "Uniform" is  (min,max,Constant), Constant default is 5. Final min and max are min/Constant and max/Constant.
In the "Normal"  is  (Mu, Sigma,Constant) , Constant default is 5. Final sd is sd/Constant.}

\item{Distribution2}{The Distribution name. ("Exponential", "Uniform" and "Normal"), if type is "Overlap".}

\item{type}{The type of splitting, "Disjoint" or "Overlap".}
}
\value{
It returns a list:

num: the number of splits.

Rs : the random number.

lr : the minimum value of the split.

ur : the maximum value of the split.
}
\description{
It maps the functional data by the random splits with the summary statistics.
}
\examples{

}
\references{
\itemize{
 \item \code{1-} : Möller, A., Tutz, G., & Gertheiss, J. (2016). Random forests for functional covariates. Journal of Chemometrics, 30(12), 715-725.
}
}
