% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForestRandomRegression.R
\docType{class}
\name{ForestRandomRegression-class}
\alias{ForestRandomRegression-class}
\alias{ForestRandomRegression}
\title{Regression forest class}
\description{
Subclass for regression forest.
Contains all fields and methods used special for random-regression forests.
}

