% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSRF_Version_1_0.R
\name{makeSIMData}
\alias{makeSIMData}
\title{Making Simulation Data}
\usage{
makeSIMData(nSS_1, nSS_2, responseType, Seed)
}
\arguments{
\item{nSS_1}{Number of Total Samples in Group 1}

\item{nSS_2}{Number of Total Samples in Group 2}

\item{responseType}{Type of response ("Categorical","Continuous" and "Survival")}

\item{Seed}{Seed number for memory.}
}
\value{
A list with 6 elements.   The 1 to 4 are functional covariets , the 5 is the dataframe for non-functional covariates and  is the response.
}
\description{
Making Simulated Data
}
\details{
It makes three functional and two non-functional covariates. The first nSS/2 is case (green color) and second nSS/2 is contorl (grey color).
}
\examples{
 #Number of Samples is 100
  nSample_1 <- 100
  nSample_2 <- 100
  SIM_1_FUNC_DATA <- makeSIMData(nSS_1 = nSample_1,nSS_2 = nSample_2, responseType ="Categorical",Seed= 2400)
  par(mfrow=c(2,3))
  matplot(t(SIM_1_FUNC_DATA[[1]]),type="l",col=c(rep("Green",nSample_1/2),rep("Grey",nSample_1/2)),xlab = expression(t),ylab = expression(X(t)) ,main= expression(X[1](t)))
  matplot(t(SIM_1_FUNC_DATA[[2]]),type="l",col=c(rep("Green",nSample_1/2),rep("Grey",nSample_1/2)),xlab = expression(t),ylab = expression(X(t)) ,main= expression(X[2](t)))
  matplot(t(SIM_1_FUNC_DATA[[3]]),type="l",col=c(rep("Green",nSample_1/2),rep("Grey",nSample_1/2)),xlab = expression(t),ylab = expression(X(t)) ,main= expression(X[3](t)))
  matplot(t(SIM_1_FUNC_DATA[[4]]),type="l",col=c(rep("Green",nSample_1/2),rep("Grey",nSample_1/2)),xlab = expression(t),ylab = expression(X(t)) ,main= expression(X[4](t)))
  boxplot(SIM_1_FUNC_DATA[[5]][1][-(1:4),][(1:nSample_1/2)], SIM_1_FUNC_DATA[[5]][1][-(1:4),][-(1:nSample_1/2)],col=c("Green","Grey"),main=expression(X[1]))
  boxplot(SIM_1_FUNC_DATA[[5]][2][-(1:4),][(1:nSample_1/2)], SIM_1_FUNC_DATA[[5]][2][-(1:4),][-(1:nSample_1/2)],col=c("Green","Grey"),main=expression(X[2]))
  par()
}
\references{
\itemize{
 \item \code{1-} : Möller, A., Tutz, G., & Gertheiss, J. (2016). Random forests for functional covariates. Journal of Chemometrics, 30(12), 715-725.
 \item \code{2-} : Ishwaran, H., Kogalur, U. B., Blackstone, E. H., & Lauer, M. S. (2008). Random survival forests. The annals of applied statistics, 2(3), 841-860.
 \item \code{3-} : Ishwaran, H., & Lu, M. (2019). Standard errors and confidence intervals for variable importance in random forest regression, classification, and survival. Statistics in medicine, 38(4), 558-582.
}
}
