% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSRF_Version_1_0.R
\name{RS_Brier_Surv}
\alias{RS_Brier_Surv}
\title{Plot the Brier Score}
\usage{
RS_Brier_Surv(
  dataset = DS,
  method = c("KM", "RFSRC"),
  dataDivid = c("TRAIN", "TEST")
)
}
\arguments{
\item{method}{It has two options: "KM","RFSRC".}

\item{dataDivid}{It has two options `Train` and `Test`.}

\item{Dataset}{It is `RF_OUTPUT$Brier_DS` from `RS_RF_SRC_H`.}
}
\value{
The Brier score plot and the summary statistics. 
                          the mean and SD of The continuous rank probability score (CRPS) (min, Q1, median, mean, Q3 and max).
}
\description{
Plot the Brier Score for the Survival Response based on the two methods Kaplan-Meier censoring distribution (KM) and RSF censoring distribution (RFSRC).
}
\details{
The `type` is `Train` or `Test` (Train= in train dataset , 2=in test dataset)
}
\examples{
#RF_Result from previous example.
par(mfrow=c(2,2))
RS_Brier_Surv(dataset = BG_Res$Brier_DS,  method = c("KM"), dataDivid = c("TRAIN"))
RS_Brier_Surv(dataset = BG_Res$Brier_DS,  method = c("KM"), dataDivid = c("TEST"))  
RS_Brier_Surv(dataset = BG_Res$Brier_DS,  method = c("RFSRC"), dataDivid = c("TRAIN"))
RS_Brier_Surv(dataset = BG_Res$Brier_DS,  method = c("RFSRC"), dataDivid = c("TEST")) 
}
\references{
\itemize{
 \item \code{1-} : Möller, A., Tutz, G., & Gertheiss, J. (2016). Random forests for functional covariates. Journal of Chemometrics, 30(12), 715-725.
 \item \code{2-} : Ishwaran, H., Kogalur, U. B., Blackstone, E. H., & Lauer, M. S. (2008). Random survival forests. The annals of applied statistics, 2(3), 841-860.
 \item \code{3-} : Ishwaran, H., & Lu, M. (2019). Standard errors and confidence intervals for variable importance in random forest regression, classification, and survival. Statistics in medicine, 38(4), 558-582.
}
}
