% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSRF_Version_1_0.R
\name{Split_Domain}
\alias{Split_Domain}
\title{Split the Domain}
\usage{
Split_Domain(
  min,
  max,
  Params = c(0, 0, 0),
  Distribution,
  Params2 = c(0, 0, 0),
  Distribution2,
  type
)
}
\arguments{
\item{min}{The minimum value of the domain.}

\item{max}{The maximum value of the domain.}

\item{Params}{The parameters of the distribution. 
In the "Exponential" is (rate, 0,0)
In the "Uniform" is  (min,max,0)
In the "Normal"  is  (Mu, Sigma,0)}

\item{Distribution}{The Distribution name. ("Exponential", "Uniform" and "Normal").}

\item{Params2}{The parameters of the second distribution, if type is "Overlap".
In the "Exponential" is (rate, Constant,0), Constant default is 5. Final rate is rate*Constant.
In the "Uniform" is  (min,max,Constant), Constant default is 5. Final min and max are min/Constant and max/Constant.
In the "Normal"  is  (Mu, Sigma,Constant) , Constant default is 5. Final sd is sd/Constant.}

\item{Distribution2}{The Distribution name. ("Exponential", "Uniform" and "Normal"), if type is "Overlap".}

\item{type}{The type of splitting, "Disjoint" or "Overlap".}
}
\value{
It returns a list:

num: the number of splits.

Rs : the random number.

lr : the minimum value of the split.

ur : the maximum value of the split.
}
\description{
It splits the domain (interval) from  minimum to maximum by the random numbers generated from Normal, Exponential or Uniform distributions.
}
\examples{
## Split a randomly a domain from 0 to 10 by random numbers from exponential distribution with rate = 1 and disjoint.
set.seed(1234)
Split_Domain(min=0,max=10, Params=c(1,0,0), Distribution = "Exponential",type="Disjoint")
## Split a randomly a domain from 0 to 10 by random numbers from exponential distribution with rate = 1 and Overlap. The the overlap is also obtained from random numbers from the Exponential distribution with the rate parameter equal to 1 and divided by the constant equal to 5  .
set.seed(1234)
Split_Domain(min=0,max=10, Params=c(1,0,0), Distribution = "Exponential",Params2=c(1,5,0),Distribution2="Exponential" , type="Overlap")
}
\references{
\itemize{
 \item \code{1-} : Möller, A., Tutz, G., & Gertheiss, J. (2016). Random forests for functional covariates. Journal of Chemometrics, 30(12), 715-725.
}
}
