% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeRandomRegression.R
\docType{class}
\name{TreeRandomRegression-class}
\alias{TreeRandomRegression-class}
\alias{TreeRandomRegression}
\title{Regression tree class}
\description{
Subclass for regression tree.
Contains all fields and methods used special for regression trees.
}

